modify calculate function

+++ ChainNode.cpp
@@ -1,6 +1,6 @@
 #include "ChainNode.h"
 
-ChainNode::ChainNode(int e, ChainNode *p)
+ChainNode::ChainNode(float e, ChainNode *p)
 {
     element = e;
     next = p;
